# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API REST de prueba",
    "description": "Esta es una API para probar la funcionalidad de generar automáticamente el documento JSON de Swagger",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://test-swagger-cibertec.herokuapp.com"
    }
  ],
  "paths": {
    "/existe-paciente": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "dni",
            "in": "body",
            "description": "DNI del paciente",
            "schema": {
              "type": "object",
              "properties": {
                "dni": {
                  "type": "string",
                  "example": "72552743"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna el paciente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/paciente": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "nombre",
            "in": "body",
            "description": "Nombre del paciente",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Peralta Westreicher"
                }
              }
            }
          },
          {
            "name": "apellidos",
            "in": "body",
            "description": "Apellidos del paciente",
            "schema": {
              "type": "object",
              "properties": {
                "apellidos": {
                  "type": "string",
                  "example": "Peralta Westreicher"
                }
              }
            }
          },
          {
            "name": "correo",
            "in": "body",
            "description": "Correo del paciente",
            "schema": {
              "type": "object",
              "properties": {
                "correo": {
                  "type": "string",
                  "example": "mario@gmail.com"
                }
              }
            }
          },
          {
            "name": "telefono",
            "in": "body",
            "description": "Telefono del paciente",
            "schema": {
              "type": "object",
              "properties": {
                "telefono": {
                  "type": "string",
                  "example": "972914057"
                }
              }
            }
          },
          {
            "name": "dni",
            "in": "body",
            "description": "DNI del paciente",
            "schema": {
              "type": "object",
              "properties": {
                "dni": {
                  "type": "string",
                  "example": "72552743"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta del registro de paciente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacientePostResp"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "dni": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/paciente/reset": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/especialidad": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/turnos": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/cita": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "especialidad",
            "in": "body",
            "description": "Especialidad para la cita",
            "schema": {
              "type": "object",
              "properties": {
                "especialidad": {
                  "type": "string",
                  "example": "psicología"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/existe-fake-reniec": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "dni",
            "in": "body",
            "description": "DNI de la persona",
            "schema": {
              "type": "object",
              "properties": {
                "dni": {
                  "type": "string",
                  "example": "72552743"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna la persona encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FakeReniec"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/fake-reniec": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Paciente": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "dni": {
            "type": "string"
          }
        }
      },
      "PacientePostResp": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FakeReniec": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          }
        }
      }
    }
  }
}